Move MultiFileReaderFunction to separate file
